Abstract USB functions from commandProcessor (still static, but progress)
